Start scanning the target's WAL from last common checkpoint.
We only need to find out about pages that were modified in the target
after the point of divergence, but hint bit updates on a page are WAL
logged only the first time after a checkpoint. So we have to assume
that any page that has had its hint bits updated after the last
checkpoint has possibly also been updated after the point of divergence.

(All this relies on page checksums to be enabled. Otherwise hint bit
updates are not logged at all.)

Pointed out by Pavan Deolasee on pgsql-hackers.