Small changes to the way the regression tests are launched
Running the regression tests required a server to be running, because
pg_regress launches the test scripts using psql against the running server.
But that server wasn't actually used for anything, the test driver initdb's
a fresh installation and does all the interesting stuff with that.
This patch removes the requirement for an existing server.

pg_regress normally launches the test scripts using psql, but launching
psql requires a server to be running, so use a custom launcher script that
runs the .sql files using a shell instead.

Also, use ./pg_rewind instead of just pg_rewind in the test script. The
current directory isn't usually in $PATH, and presumably we want to test the
version that's compiled.