Remove local directories non-existing on remote side when rewinding
The fetching interface already had the tools to remove files but actually
it lacked support for directories that could exist locally but not
remotely when a database is created, removed, or when simply some temporary
folders are created like "pgsql_tmp" for on-disk sort operations. Those
temporary folders and files should actually be ignored when rewinding, but
this is food for another patch.

Bug reported by Samrat Revagade, though I didn't use his patch.