Fix assertion failure if timeline switch happened at page boundary.
If the master and standby diverged at a page boundary, we would start the
scan for the last checkpoint record at the page boundary. XLogReadRecord
is not cool with that; it expects there to be a valid WAL record at the
start point. To fix, skip over the page header.