Fix rewound node indefinitely waiting for XLOG_BACKUP_END
This behavior is findable when for example enforcing a checkpoint
on a master node once a standby has connected. In order to fix that,
control file of the rewound node is updated to make it restart
properly from archive recovery, ready to catch up fresh content
from either streaming replication of WAL archives.

At the same time, update regression tests to cover the problem that
was reported by making the rewound master connecting with streaming
replication to the newly-promoted node (which is the old standby)
and by kicking a checkpoint after standby has been stated when cluster
is initialized in tests. I didn't think that it was worth adding a
new test just for that by seeing the similarities with basictest.

Report and basic patch by Samrat Revagade, heavily edited by myself.