Skip ephemeral files of pg_xlog from source server
When fetching files from source server, if the timing is rather bad it
is possible to get at the same time temporary files of pg_xlog, files that
get normally quickly removed by the backend. This results in a error when
trying to fetch chunks from a file that does not exist.

Per report and analysis from dev1ant, patch from me.