Add module pg_rewind_support
This module contains a set of C functions that will be used by pg_rewind
to ensure that no failures are encountered when rewinding from a source
node using the stream mode. The functions added are the following ones
with basically a missing_ok option to ignore any unexistent file entries:
- ls_dir
- stat_file
- read_binary_file
Note that those functions are available in Postgres upstream in 9.5 with
pg_rewind itself.