Prevent errors when files are deleted on remote server
In stream mode where the remote server is running, it is possible that
some files that were included in the remote file list have been removed
when fetching their chunks. Use the new function set in the support
bundle included with this tool with the missing_ok option to not error
out when a file is missing but simply to ignore them.