Fix several memory leaks
A couple of calls to PQclear forgot to reset the memory of results
generated by libpq and more importantly, the regex expression used
to check if a given path is a relation file or not was recompiled
all the time, leading to a huge memory consumption.

Per report from Vladimir Borodin (dev1ant), who also provided the
run of valgrind showing the leaks.