Fix handling of pg_xlog when defined as link on target node
Deployments of Postgres instances using pg_xlog as a soft link to a
different path, be it usually a different partition, failed to work with
pg_rewind that was unabled to handle its content correctly. Fix this by
assuming that pg_xlog is a directory instead of a symlink, which is
similar to the handling done by 0e42397 in upstream code.

Add as well a test case using the new fork point introduced previously.