Don't error out when two nodes are on the same timeline
pg_rewind is taught as well to check that even a node that has been
shutdown in recovery is safe to use as a target node. The source node
gains the same checks when the local mode is activated.

This is similar to PostgreSQL upstream commit 4b021aa6, and this change
is aimed at putting the older versions of pg_rewind in line to what is
done upstream to avoid a jump in use.