Fix minimum recovery targets after rewind
pg_rewind enforces minimum recovery target on the target node at the end
of processing. However, up to now the minimum recovery timeline has been
set as the timeline of the last checkpoint that happened on target node,
and the minimum recovery point at the divergence point, which is where
WAL forked. However, this proves to be unreliable, because a rewound node
entering in recovery can consistently replay only up to the LSN point of
source node after all the files modified have been copied to the target.
This commit does that, in the case of an online source node the current
LSN insert position is used as minimum recovery point. This will avoid
also things like pg_isready to report atoo-early success and potential
read-only queries to hit a hot standby in recovery where they could be
inconsistent pages.