Add set of sanity checks when using --source-server
When establishing a connection to a source server to use it as a source
for rewinding a target, a couple of precautions need to be taken:
- When source server is set up for synchronous replication, if it is for
a reason or another pg_rewind can remain stuck for no real reason, so
enforce synchronous_commit to off when establishing connection as a set
of objects are created. This comes from upstream commit d7eb76b9.
- Ensure that the source server is not in recovery. As a set of objects
is created to make the operation work this would fail afterwards.
- Check that full_page_writes are enabled, as a precaution against torn
pages when reading them.

The second and third problems have been noticed while looking at the first
one which has just been committed upstream.