Improve pg_rewind handling of relation data files in tablespaces
This is a rough backpatch of upstream commit cb5c141e which is using a
sscanf method instead of a regex like this version of pg_rewind to
identify what are relation data files.  In order to make future
back-patches more easy and the code more portable, I have decided to
copy the upstream logic into this tree.