Empty search_path when running pg_rewind
This makes pg_rewind behave as documented by upstream PostgreSQL
regardless of the connect-time search_path and regardless of
user-created objects.  Today, a malicious user with CREATE permission on
a search_path schema can take control of certain of these clients'
queries and invoke arbitrary SQL functions under the client identity,
often a superuser.  This is exploitable in the default configuration,
where all users have CREATE privilege on schema "public".

Note that pg_rewind_support uses its own schema at installation (we
could as well enforce the extension to not be moveable across schemas),
and that those functions are already qualified, so things are safe.

CVE-2018-1058