Make regression tests of pg_rewind more stable on slow machines
The tests have historically been using a set of hardcoded sleeps to wait
for a standby initially started to catch up with its primary and for a
promotion.  On slow machines, this is proving to be unreliable as it may
take more than 1s to actually finish those steps.  Instead of the
hardcoded sleeps, query directly PostgreSQL to make sure that the wanted
state is reached.

Backpatch down to 9.3, where those tests are used.

Testing Done:
- Regression tests are able to pass properly (added some custom logs to
make sure that the results are correct, which are not part of this final
patch as they generate regression diffs).